+++ service/server/views/game/duet.py
@@ -1,6 +1,5 @@
 # pylint: disable=R0801
 
-import requests
 import ulid
 from codenames.duet.board import DuetBoard
 from codenames.duet.state import DuetGameState
@@ -8,11 +7,9 @@ from codenames.generic.move import Clue, Guess
 from codenames.utils.vocabulary.languages import get_vocabulary
 from rest_framework.viewsets import GenericViewSet
 from the_spymaster_api.structs import (
-    BaseRequest,
     ClueRequest,
     GetGameStateRequest,
     GuessRequest,
-    HttpResponse,
     NextMoveRequest,
 )
 from the_spymaster_api.structs.duet.requests import StartGameRequest
@@ -23,13 +20,10 @@ from the_spymaster_api.structs.duet.responses import (
     NextMoveResponse,
     StartGameResponse,
 )
-from the_spymaster_solvers_api.structs.requests import LoadModelsRequest
-from the_spymaster_solvers_api.structs.responses import LoadModelsResponse
 from the_spymaster_util.http.errors import BadRequestError
 from the_spymaster_util.logger import get_logger
 
 from server.logic.db import get_or_create, save_game
-from server.logic.solvers import get_solvers_client
 from server.models.game import DuetGame
 from server.views.endpoint import HttpMethod, endpoint
 
@@ -41,9 +35,6 @@ def ulid_lower():
 
 
 class DuetGameView(GenericViewSet):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.solvers_client = get_solvers_client()
 
     @endpoint
     def start(self, request: StartGameRequest) -> StartGameResponse:
@@ -95,27 +86,3 @@ class DuetGameView(GenericViewSet):
         # game.state_data = handler.game_state.model_dump()
         # save_game(game)
         # return response
-
-    @endpoint(url_path="load-models")
-    def load_models(self, request: LoadModelsRequest) -> LoadModelsResponse:
-        response = self.solvers_client.load_models(request)
-        return response
-
-    @endpoint(methods=[HttpMethod.GET])
-    def test(self, request: BaseRequest) -> HttpResponse:  # pylint: disable=unused-argument
-        body = {"details": "It seems everything is working!"}
-        return HttpResponse(body=body)
-
-    @endpoint(methods=[HttpMethod.GET], url_path="raise-error")
-    def raise_error(self, request: BaseRequest) -> HttpResponse:
-        if getattr(request, "handled", False):
-            raise BadRequestError(message="This error should be handled!")
-        if getattr(request, "good", False):
-            return HttpResponse(body={"success": "true"})
-        raise Exception("Test error")  # pylint: disable=broad-exception-raised
-
-    @endpoint(methods=[HttpMethod.GET], url_path="ping-google")
-    def ping_google(self, request: BaseRequest) -> HttpResponse:  # pylint: disable=unused-argument
-        response = requests.get("https://www.google.com", timeout=10)
-        body = {"status_code": response.status_code, "duration": response.elapsed.total_seconds()}
-        return HttpResponse(body=body)
